Extract album cover retrieval
